cuda port visual studio build correction

+++ nrgcore/mineplant.cpp
@@ -168,7 +168,7 @@ void MinePlant::collectHashRate()
     for (auto const& miner : m_miners) {
         auto minerHashCount = miner->hashCount();
         p.hashes += minerHashCount;
-        p.minersHashes.insert(std::make_pair<std::string, uint64_t>(miner->name(), minerHashCount));
+        p.minersHashes.insert(std::make_pair<std::string, uint64_t>(miner->name(), uint64_t(minerHashCount)));
     }
 
     //Reset miner hashes
